backend/requirements.txt
python-dotenv
fastapi
uvicorn
google-generativeai
yfinance
